Move all template commands to new package
