+++ sernet.gs.service/src/sernet/verinice/service/commands/templates/LoadTemplatesOrImplementations.java
@@ -17,7 +17,7 @@
  * Contributors:  
  *     Viktor Schmidt <vschmidt[at]ckc[dot]de> - initial API and implementation  
  ******************************************************************************/
-package sernet.verinice.service.commands;
+package sernet.verinice.service.commands.templates;
 
 import java.io.Serializable;
 import java.util.Collections;
@@ -33,6 +33,7 @@ import sernet.verinice.interfaces.GenericCommand;
 import sernet.verinice.interfaces.IBaseDao;
 import sernet.verinice.model.common.CnATreeElement;
 import sernet.verinice.model.common.CnATreeElement.TemplateType;
+import sernet.verinice.service.commands.LoadElementsByUuid;
 
 /**
  * This command loads for given ({@link CnATreeElement}) all modeling templates
